Create http server on your Windows Desktop
(only Windows) 
In order for the script to work, it has to be converted to .exe (using "pyinstaller --onefile desktop-hidden-http.pyw").
On execution the exe file gets moved to the desktop where a local http-server on port 9999 is created. Additionally the files attribute is set to "hidden". (change your Windows File Explorer Options to show hidden files in order to see it). The server can now be reached from the browser using http://localhost:9999 or http://yourIP:9999