Python3 Reverse Shell (first try-learning purpose)
This is my first try on creating a reverse shell script on python3 (learning purposes). Run it on the target server while listening on your client using "nc -lvp 1234". Don't forget to change the ip-adress in the code.  
The shell is still unconsistent (does not execute commands with slashes "/").
Will keep developing